BZ-1316966: extend test with kie:environment defined in blueprint
(cherry picked from commit 9c2a66cf9e414fe6893a6154e580d1e93d30f09b)
