+++ drools-osgi/drools-karaf-itests/src/test/resources/org/drools/karaf/itest/blueprint/processpersistence/kie-beans-blueprint-process-persistence.xml
@@ -11,17 +11,22 @@
   <reference id="userTransactionRef" interface="javax.transaction.UserTransaction" />
   <reference id="transactionSynchronizationRegistryRef" interface="javax.transaction.TransactionSynchronizationRegistry" />
 
-  <!-- this would work if other properties (such as UserTransaction and TransactionSynchronizationRegistry) were set as well, however the XSD does not allow that -->
-  <!--
   <kie:environment id="environmentWithPersistence">
       <kie:transaction-manager ref="transactionManagerRef" />
       <kie:transaction-sync-registry ref="transactionSynchronizationRegistryRef" />
       <kie:user-transaction ref="userTransactionRef" />
       <kie:entity-manager-factory ref="entityManagerFactoryRef" />
   </kie:environment>
-   -->
 
-  <bean id="processBean" class="org.drools.karaf.itest.beans.ProcessWithPersistenceBean"
+    <!-- Configuring persistence using kie:environment -->
+  <bean id="processBeanWithEnv" class="org.drools.karaf.itest.beans.ProcessWithPersistenceEnvBean"
+        init-method="init" destroy-method="destroy">
+
+     <argument index="0" ref="environmentWithPersistence" />
+  </bean>
+
+  <!-- Configuring persistence without using kie:environment -->
+  <bean id="processBeanWithoutEnv" class="org.drools.karaf.itest.beans.ProcessWithPersistenceDirectBean"
         init-method="init" destroy-method="destroy">
 
       <argument index="0" ref="entityManagerFactoryRef" />
@@ -30,7 +35,19 @@
       <argument index="3" ref="transactionSynchronizationRegistryRef" />
   </bean>
 
-  <bean id="kieSessionBean" factory-ref="processBean" factory-method="createKieSession" destroy-method="destroy" />
-  <service id="kieSessionService" interface="org.kie.api.runtime.KieSession" ref="kieSessionBean" />
+  <!-- Exporting KIE Sessions as OSGi services for the test (PAx-Exam won't let me export processBeans themselves) -->
+  <bean id="kieSessionWithEnvBean" factory-ref="processBeanWithEnv" factory-method="createKieSession" />
+  <service id="kieSessionWithEnvService" interface="org.kie.api.runtime.KieSession" ref="kieSessionWithEnvBean" >
+    <service-properties>
+        <entry key="osgi.jndi.service.name" value="kiesession/sessionWithEnv"/>
+    </service-properties>
+  </service>
+
+  <bean id="kieSessionWithoutEnvBean" factory-ref="processBeanWithoutEnv" factory-method="createKieSession" />
+  <service id="kieSessionWithoutEnvService" interface="org.kie.api.runtime.KieSession" ref="kieSessionWithoutEnvBean" >
+    <service-properties>
+        <entry key="osgi.jndi.service.name" value="kiesession/sessionWithoutEnv"/>
+    </service-properties>
+  </service>
 
 </blueprint>
